+++ CHANGELOG.md
@@ -57,6 +57,7 @@
 * Fix #6628: Removed deprecated `io.fabric8.kubernetes.client.CustomResourceList` class
 * Fix #6630: Removed deprecated `io.fabric8.kubernetes.client.utils.ReflectUtils` class
 * Fix #6455: (crd-generator) Changed format in `@PrinterColumn` to enum
+* Fix #6673: (crd-generator) Deprecation of CRD Generator v1
 
 ### 6.13.4 (2024-09-25)